+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 5.3.0 - undefined
 
+## 6.0 - undefined
+### Added
+- Screen Sharing
+
+## 5.3.0 - undefined
 ### Fixed
+- Update SDK to 5.4
+- Urls (Privacy policy, terms, file transfer, update check)
 - Blinking on resize or Ui updates on screen that have low frame rates.
 - Timezones and Windows fix.
 - Download the correct version of openH264 (2.2.0)